requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
Werkzeug==2.2.3
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
email-validator==2.0.0
Flask-Mail==0.9.1
Pillow==9.5.0
pywhatkit==5.4
gunicorn==20.1.0
